fix(openiddict): move SigningKey to Domain/ folder and fix Wolverine test isolation
Move SigningKey.cs from Entities/ to Domain/ to satisfy
FileOrganizationTests (domain types must reside in Domain/ subfolder).

Add [Collection("TraceContext")] to TraceContextBehaviorTests and
TraceContextBehaviorAdditionalTests to prevent parallel execution —
both classes share a global ActivityListener on the same ActivitySource,
causing cross-test activity capture.
