+++ tests/Granit.Wolverine.Tests/TraceContextBehaviorAdditionalTests.cs
@@ -16,6 +16,7 @@ using Xunit;
 
 namespace Granit.Wolverine.Tests;
 
+[Collection("TraceContext")]
 public sealed class TraceContextBehaviorAdditionalTests : IDisposable
 {
     private const string ValidTraceParent = "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01";
